@@ -1,8 +1,12 @@
-const fs = require("fs-extra");
-const path = require("path");
-const { execSync } = require("child_process");
-const chalk = require("chalk");
-const ora = require("ora");
+import fs from "fs-extra";
+import path from "path";
+import { fileURLToPath } from "url";
+import { execSync } from "child_process";
+import chalk from "chalk";
+import ora from "ora";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 // 可用的模板
 const TEMPLATES = {
@@ -43,7 +47,6 @@ async function createProject(projectName, options) {
     await copyTemplate(templateName, projectName);
     await updateProjectName(projectName);
     await installDependencies(projectName);
-    await initGit(projectName);
     showSuccess(projectName, template);
   } catch (error) {
     // 清理失敗的建立
@@ -113,25 +116,6 @@ async function installDependencies(projectName) {
   }
 }
 
-async function initGit(projectName) {
-  const spinner = ora("🔧 Initializing Git repository...").start();
-
-  try {
-    const projectPath = path.resolve(projectName);
-
-    execSync("git init", { cwd: projectPath, stdio: "pipe" });
-    execSync("git add .", { cwd: projectPath, stdio: "pipe" });
-    execSync('git commit -m "Initial commit from vue-tailwind-starter"', {
-      cwd: projectPath,
-      stdio: "pipe",
-    });
-
-    spinner.succeed("✅ Git repository initialized");
-  } catch (error) {
-    spinner.warn("⚠️  Could not initialize Git repository (this is optional)");
-  }
-}
-
 function detectPackageManager() {
   try {
     execSync("pnpm --version", { stdio: "pipe" });
@@ -165,6 +149,4 @@ function showSuccess(projectName, template) {
   console.log("");
 }
 
-module.exports = {
-  createProject,
-};
+export { createProject };
